moved test_X_data to conftest.py
