+++ tests/test_data_load_and_split.py
@@ -4,23 +4,6 @@ import pandas as pd
 import numpy as np
 import pytest
 
-# df = factory_data_load_and_split.load_multilabel_data(filename = 'datasets/testing/test_data.csv', target = 'major_categories')[:500]
-# X_train_val, X_test, Y_train_val, Y_test = factory_data_load_and_split.process_and_split_multilabel_data(df, target = major_cats, preprocess_text = False, additional_features = True)
-
-@pytest.fixture
-def grab_test_X_additional_feats():
-    data_dict = {'FFT answer': {'Q1': 'Nurses were great',
-                    'Q2': 'Communication was fantastic',
-                    'Q3': 'Impossible to find parking, but pleased to get an appointment close to home',
-                    'Q4': 'Food and drink selection very limited',
-                    'Q5': 'The ward was boiling hot, although staff were great at explaining details'},
-                    'FFT_q_standardised': {'Q1': 'what_good',
-                    'Q2': 'what_good',
-                    'Q3': 'could_improve',
-                    'Q4': 'could_improve',
-                    'Q5': 'could_improve'}}
-    text_X_additional_feats = pd.DataFrame(data_dict)
-    return text_X_additional_feats
 
 @pytest.fixture
 def grab_test_Y():
